Set the title in the navigation bar from the web view title
The `VisitableViewController` has code to do this for UIKit, but it
doesn't work for SwiftUI because it wraps a view around the
`UIViewControllerRepresentable` view object. We must set it on that
wrapper which means setting it on the parent. I created my own class to
do this.

One bug with this solution: when you click the "Back" navigation item,
the title correctly changes for half a second and then disappears. After
playing around with a lot of things, I've determined it happens after
`viewDidAppear` for the `UIViewControllerRepresentable` (`TurboView`) is
called. I think this means there's nothing I can do to fix this bug of
the title being overridden. (Note: This doesn't happen if I click a link
that takes me back, which means we can override the back button with our
own implementation.)