Make TurboView a SwiftUI view wrapping the UIKit view controller
Now `TurboView` is a native SwiftUI view that wraps `TurboUIWrapperView`
which is the `UIViewControllerRepresentable` that contains the
`VisitableViewController`. This is all done so that the `TurboView`
wrapper can contain the state for the navigation title and allow the
`TurboUIWrapperView` to take a binding to it to update the value. In
other words, it means we can easily set the navigation toolbar title to
the title of the web page after it loads, and SwiftUI will persist it,
even when pushing the back button.