Make NavStack its own thing
This creates `NavStack` which allows us to easily navigate back to a
view if an entry already exists. Here's what I mean, if you navigate to
view A and then B and then C, your stack looks like `[A, B, C]`. If you
then try to advance to view B again, we don't want the stack to become
`[A, B, C, B]`. Instead, we pop off the elements until we get back to B:
`[A, B]`.

This functionality was tied up in TurboNavHelper, but is useful as its
own thing and can be used for other navigation stacks that don't include
Turbo views.