Fix bugs around the root view
Because the root view isn't part of the stack / array, it isn't always
properly updated when modals are dismissed, and, before this commit,
could actually end up as one of the views on the stack. This fixes that
by tracking the root proposal so that we can navigate back to it instead
of adding it forward on the stack. If the stack is empty when we
"navigate" to it, we also reload because it doesn't change to the new
HTML. (Even with `@Published`, the new HTML I see in the `proposal` does
not get loaded.)